LayoutTests/fast/scrolling/scroll-into-view-block-center.html

<!DOCTYPE html>
<html>
<head>
    <style>
        .scroller {
            width: 300px;
            height: 300px; /* Chosen to cause the available height to be odd with 15px scrollbars */
            overflow: scroll;
            border: 1px solid black;
        }
        
        #target {
            width: 100%;
            height: 30px;
        }
        
        .spacer {
            width: 10px;
            height: 1000px;
            background-color: silver;
        }
    </style>
    <script src="../../resources/js-test-pre.js"></script>
    <script>
        let scroller;
        let scrollTopAfterFirstScrollIntoView;
        window.addEventListener('load', () => {
            let target = document.getElementById('target');
            scroller = document.getElementsByClassName('scroller')[0];

            target.scrollIntoView({ block: 'center' });
            scrollTopAfterFirstScrollIntoView = scroller.scrollTop;

            target.scrollIntoView({ block: 'center' });
            shouldBe('scroller.scrollTop', 'scrollTopAfterFirstScrollIntoView');
        }, false);
    </script>
</head>
<body>
    <div class="scroller">
        <div class="spacer"></div>
        <div id="target">This should not jiggle</div>
        <div class="spacer"></div>
    </div>
    <div id="console"></div>
    <script src="../../resources/js-test-post.js"></script>
</body>
</html>
