LayoutTests/fast/scrolling/scroll-into-view-block-center-expected.txt

This should not jiggle
PASS successfullyParsed is true

TEST COMPLETE
PASS scroller.scrollTop is scrollTopAfterFirstScrollIntoView
